Evolution: Reduce build order search space
Let only children be surviors for next generation

Move individual generation/transformation, selection
and evaluation to own directories.
