+++ src/coevolution/evaluation/Evaluation.js
@@ -1,0 +1,19 @@
+import Game from '~/rts/Game';
+const maxGameLoops = 400;
+
+export function evaluate_wip(individuals) {
+    individuals.forEach(individual => {
+        individual.fitness = 0;
+    });
+
+    individuals.forEach(one => {
+        individuals.filter(two => two !== one).forEach(two => {
+            const game = new Game('unnecessary-id', maxGameLoops, one, two);
+
+            game.simulate();
+
+            one.fitness += game.finalScore.blue.score;
+            two.fitness += game.finalScore.red.score;
+        });
+    });
+}
