+++ src/coevolution/selection/index.js
@@ -1,0 +1,3 @@
+export {selectUnique} from './selectUnique';
+export {select as rouletteWheelSelection} from './rouletteWheelSelection';
+export {select as linearRankSelection} from './linearRankSelection';
